+++ src/components/Comment/CommentInputForm/CommentInputForm.module.scss
@@ -1,0 +1,17 @@
+.commentInputForm{  
+  margin: 1rem 1.5rem;
+  textarea{
+    width: 100%;
+    background-color: var(--bg-primary);
+    color: var(--text-primary);
+    resize: none;
+    padding: 1rem 1rem;
+    outline: none;
+    border: 1px solid var(--text-secondary);
+    margin-bottom: 1.5rem;
+    border-radius: 4px;
+    min-height: 6.125rem;
+    font-size: 1rem;
+    line-height: 1.75;
+  }
+}
