+++ src/components/Comment/CommentInputForm/index.tsx
@@ -3,7 +3,7 @@ import styles from './CommentInputForm.module.scss';
 const index = () => {
   return (
     <div className={styles.commentInputForm}>
-      <textarea placeholder='댓글을 입력해주세요.' />
+      <textarea placeholder='댓글 부탁해요.' />
       <div className={styles.buttonWrapper}></div>
     </div>
   );
